pyproject.toml
# see https://flit.pypa.io/en/latest/pyproject_toml.html
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nicegraph"
description = "nicegraph: Graph Navigation User Interface"
home-page = "https://github.com/BITPlan/nicegraph"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" },
]
readme = "README.md"
license = {text = "Apache-2.0"}
dependencies = [
	# https://pypi.org/project/pybasemkit/
  	# https://github.com/WolfgangFahl/pybasemkit
  	# Python base module kit: YAML/JSON I/O, structured logging, CLI tooling, shell execution, and remote pydevd debug support.
 	"pybasemkit>=0.0.5",
    # https://github.com/WolfgangFahl/nicegui_widgets
    "ngwidgets>=0.29.1",
    # https://pypi.org/project/pyLodStorage/
	"pyLodStorage>=0.17.0",
    #https://pypi.org/project/pydevd/
    'pydevd',
    # https://pypi.org/project/rdflib/
    'rdflib>=7.0.0',
    # https://pypi.org/project/pyMogwai/
    "pyMogwai>=0.0.8",
    # internationalization
    # https://github.com/danhper/python-i18n
    "python-i18n>=0.3.9"
 ]

requires-python = ">=3.10"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: User Interfaces",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
    ]
dynamic = ["version"]
[tool.hatch.version]
path = "ng3/__init__.py"

[project.urls]
Home = "https://github.com/BITPlan/nicegraph"
Documentation = "https://wiki.bitplan.com/index.php/nicegraph"
Source = "https://github.com/BITPlan/nicegraph"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["ng3"]

[tool.hatch.build.targets.wheel.sources]
"ng3" = "ng3"


[project.scripts]
ng3 = "ng3.ng3_cmd:main"
